requirements.txt
streamlit
pillow
matplotlib
roboticstoolbox-python
opencv-contrib-python
scikit-learn
pyserial
numpy<2